requirements.txt
Django==1.10.5
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.7
